added staffview database

+++ src/main/kotlin/com/example/demo/model/StaffCategory.kt
@@ -1,6 +1,7 @@
 package com.example.demo.model
 
 import javafx.beans.property.SimpleDoubleProperty
+import javafx.beans.property.SimpleIntegerProperty
 import javafx.beans.property.SimpleStringProperty
 import tornadofx.*
 
@@ -18,7 +19,7 @@ class StaffCategory(val cat_id: Int = -1){
     val expReqProperty = SimpleStringProperty(this, "expReq", "4 Years")
     var expReq by expReqProperty
 
-    val baseSalaryProperty = SimpleDoubleProperty(this, "Base Salary", 0.0)
+    val baseSalaryProperty = SimpleIntegerProperty(this, "Base Salary", 0)
     var baseSalary by baseSalaryProperty
 
     override fun toString(): String {
